réaliser la méthode ajouter ticket

+++ Backend/demo2/src/main/java/com/example/demo2/dao/PersonneRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface PersonneRepository extends JpaRepository <Personne, Integer> {
+public interface PersonneRepository extends JpaRepository <Personne, Long> {
 
     Optional<Personne> findByEmail(String email);
 }
